Fix repeat options being focusable when hidden
This occurs because of a bug in GTK3 which doesn't prevent focus when
widgets are hidden. To work around this, the 'Can focus' property in
buttons under the repeat expander is by default off to prevent them from
having focus while cycling through the options. A signal is emitted
when the user clicks on the expander to toggle them.

Closes #233